test for SHOW_NEXT_COORDS
