+++ src/reducers/game.ts
@@ -1,30 +1,7 @@
 import { Action } from '../actions';
-import { GameState, Coordinate } from '../types';
-import { files, ranks } from '../constants/models';
+import { GameState } from '../types';
 import { SHOW_NEXT_COORDS, HANDLE_SQUARE_CLICKED } from '../constants/actions';
-
-/**
- * Returns a random integer between min (inclusive) and max (inclusive)
- */
-function getRandomInt(min: number, max: number) {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-function generateNextCoord<T>(possibleCoords: T[], currentCoord?: T): T {
-  const maxCoordIndex = possibleCoords.length - 1;
-  let nextCoord = possibleCoords[getRandomInt(0, maxCoordIndex)];
-  while (nextCoord === currentCoord) {
-    nextCoord = possibleCoords[getRandomInt(0, maxCoordIndex)];
-  }
-  return nextCoord;
-}
-
-function generateRandomCoords(lastCoordinates?: Coordinate): Coordinate {
-  const nextRank = generateNextCoord(ranks, lastCoordinates ? lastCoordinates.rank : undefined);
-  const nextFile = generateNextCoord(files, lastCoordinates ? lastCoordinates.file : undefined);
-
-  return { rank: nextRank, file: nextFile };
-}
+import { generateRandomCoords } from '../lib/boardUtils';
 
 export default function game(
   state: GameState = {
